Fix issue where posts with only links didn't extract the link
This closes #5. Posts on the Hacker News RSS feed only publish a link to
the comments in the description field, so we have to check if the entire
parsed HTML is already a link tag to be able to include that link before
searching for links in the root element's children.