Use async/await to download feed contents simultaneously
It turns out that most of the bottleneck in loading the feeds is in
parsing them. This cannot be done in parallel yet because of the
external dependency on html2text.py. I'm going to prioritize
nim-html2text (https://git.sr.ht/~reesmichael1/nim-html2text) now so
that this dependency can be removed and the feeds can be parsed in
parallel.